+++ NEWS.md
@@ -15,6 +15,7 @@ by using the reference semantic of the data.table package. If many columns are a
 - The argument `subset` for the `subset` method for `remote_corpus` objects can now be a call (#162), this is a basis for passing vectors to OpenCPU server.
 - `p_attributes()` implemented for `remote_corpus` and `remote_partition`.
 - A new `regions()` method (for `corpus` class objects to start with) returns a `regions` class object with a regions matrix (slot `cpos`) with regions for an s-attribute (#176).
+- The `get_token_stream()`-method for `regions` and `matrix` objects will now accept a logical argument `split`. If `TRUE`, a list of character vectors is returned. The envisaged use case is a fast decoding of sentences (#176).
 
 ## Minor Improvements